+++ desktop-widgets/modeldelegates.cpp
@@ -15,8 +15,8 @@
 #include "qt-models/diveplannermodel.h"
 #include "qt-models/divetripmodel.h"
 #include "qt-models/divelocationmodel.h"
-#include "core/qthelper.h"
 #include "core/divesite.h"
+#include "core/pref.h"
 #include "core/selection.h"
 #include "core/string-format.h"
 #include "desktop-widgets/simplewidgets.h"
