+++ map-widget/qmlmapwidgethelper.cpp
@@ -11,6 +11,7 @@
 #include "core/divesite.h"
 #include "core/qthelper.h"
 #include "core/range.h"
+#include "core/string-format.h"
 #include "qt-models/maplocationmodel.h"
 #include "qt-models/divelocationmodel.h"
 #ifndef SUBSURFACE_MOBILE
