+++ qt-models/divesiteimportmodel.cpp
@@ -1,6 +1,5 @@
 #include "divesiteimportmodel.h"
 #include "core/divelog.h"
-#include "core/qthelper.h"
 #include "core/range.h"
 #include "core/string-format.h"
 #include "core/taxonomy.h"
